Improve MemberDAO interface, annotate InMemoryMemberDAO methods
The MemberDAO interface has been significantly expanded to include additional query operations. All methods in the InMemoryMemberDAO class that implement the MemberDAO interface have been annotated with @Override for better code clarity and maintainability. Also, the type of the 'dao' object in InMemoryMemberDAOTests has been changed to MemberDAO to better stick to programming to an interface, not an implementation.