Add sorting methods for Golfer class
Added a new feature in the sorting module that allows for different ways to sort objects of Golfer class. This was done to provide various sorting options for users like by score, by last name and by first name in both ascending and descending orders. Additionally, a partitionByScore method was added to create a map for grouping golfers based on given specifier (score < 70).