Add ManagerTest and update Department class
A new test class, ManagerTest, has been added to improve code coverage. In DepartmentTest and Department classes, changed the way Manager is associated with the Department. Instead of using the Manager as a constructor argument, a setter method has been introduced to make the association. This change allows a Department to exist without a Manager and gives us the flexibility to assign or change the manager later.